Refactor MonitorDialog and update schemas for enhanced functionality
- Deleted the MonitorDialog component to streamline the codebase and improve maintainability.
- Updated monitorSchemas to include additional fields for HTTP and TCP configurations, enhancing validation and flexibility.
- Adjusted import paths for MonitorDialog in relevant pages to reflect the new structure, ensuring consistent usage across the application.
