+++ client/src/app/uptime/components/dialog/GeneralTab.tsx
@@ -1,0 +1,234 @@
+import React from "react";
+import { UseFormReturn } from "react-hook-form";
+import { Input } from "@/components/ui/input";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { Switch } from "@/components/ui/switch";
+import { FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
+import { UptimeMonitor } from "@/api/uptime/monitors";
+
+interface GeneralTabProps {
+  form: UseFormReturn<any>;
+  monitor?: UptimeMonitor;
+  isEdit: boolean;
+  monitorType: "http" | "tcp";
+}
+
+const INTERVAL_OPTIONS = [
+  { value: 60, label: "1 minute" },
+  { value: 300, label: "5 minutes" },
+  { value: 600, label: "10 minutes" },
+  { value: 1800, label: "30 minutes" },
+  { value: 3600, label: "1 hour" },
+  { value: 7200, label: "2 hours" },
+  { value: 21600, label: "6 hours" },
+  { value: 86400, label: "24 hours" },
+];
+
+const HTTP_METHODS = ["GET", "POST", "PUT", "DELETE", "HEAD", "OPTIONS", "PATCH"] as const;
+
+export function GeneralTab({ form, monitor, isEdit, monitorType }: GeneralTabProps) {
+  return (
+    <div className="space-y-4">
+      {/* Monitor Name */}
+      <FormField
+        control={form.control}
+        name="name"
+        render={({ field }) => (
+          <FormItem>
+            <FormLabel>Monitor Name</FormLabel>
+            <FormControl>
+              <Input placeholder="My API Endpoint" {...field} />
+            </FormControl>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+
+      {/* Monitor Type */}
+      {!isEdit ? (
+        <FormField
+          control={form.control}
+          name="monitorType"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Monitor Type</FormLabel>
+              <Select onValueChange={field.onChange} defaultValue={field.value}>
+                <FormControl>
+                  <SelectTrigger>
+                    <SelectValue />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  <SelectItem value="http">HTTP/HTTPS</SelectItem>
+                  <SelectItem value="tcp">TCP Port</SelectItem>
+                </SelectContent>
+              </Select>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+      ) : (
+        <div className="space-y-1">
+          <FormLabel>Monitor Type</FormLabel>
+          <div className="px-3 py-2 border border-neutral-800 rounded-md bg-neutral-900 text-neutral-400">
+            {monitor?.monitorType === "http" ? "HTTP/HTTPS" : "TCP Port"}
+          </div>
+        </div>
+      )}
+
+      {/* HTTP Configuration */}
+      {monitorType === "http" && (
+        <>
+          <FormField
+            control={form.control}
+            name="httpConfig.url"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>URL</FormLabel>
+                <FormControl>
+                  <Input type="url" placeholder="https://api.example.com/health" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <div className="grid grid-cols-2 gap-4">
+            <FormField
+              control={form.control}
+              name="httpConfig.method"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>HTTP Method</FormLabel>
+                  <Select onValueChange={field.onChange} value={field.value} defaultValue={field.value}>
+                    <FormControl>
+                      <SelectTrigger>
+                        <SelectValue />
+                      </SelectTrigger>
+                    </FormControl>
+                    <SelectContent>
+                      {HTTP_METHODS.map((method) => (
+                        <SelectItem key={method} value={method}>
+                          {method}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+
+            <FormField
+              control={form.control}
+              name="httpConfig.timeoutMs"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Timeout (ms)</FormLabel>
+                  <FormControl>
+                    <Input
+                      type="number"
+                      {...field}
+                      value={field.value || ""}
+                      onChange={(e) => field.onChange(e.target.value ? parseInt(e.target.value) : undefined)}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+          </div>
+        </>
+      )}
+
+      {/* TCP Configuration */}
+      {monitorType === "tcp" && (
+        <div className="grid grid-cols-2 gap-4">
+          <FormField
+            control={form.control}
+            name="tcpConfig.host"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Host</FormLabel>
+                <FormControl>
+                  <Input placeholder="example.com" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="tcpConfig.port"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Port</FormLabel>
+                <FormControl>
+                  <Input
+                    type="number"
+                    {...field}
+                    value={field.value || ""}
+                    onChange={(e) => field.onChange(e.target.value ? parseInt(e.target.value) : undefined)}
+                  />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+      )}
+
+      {/* Check Interval */}
+      <FormField
+        control={form.control}
+        name="intervalSeconds"
+        render={({ field }) => (
+          <FormItem>
+            <FormLabel>Check Interval</FormLabel>
+            <Select
+              onValueChange={(value) => field.onChange(parseInt(value))}
+              value={field.value?.toString()}
+              defaultValue={field.value?.toString()}
+            >
+              <FormControl>
+                <SelectTrigger>
+                  <SelectValue />
+                </SelectTrigger>
+              </FormControl>
+              <SelectContent>
+                {INTERVAL_OPTIONS.map((option) => (
+                  <SelectItem key={option.value} value={option.value.toString()}>
+                    {option.label}
+                  </SelectItem>
+                ))}
+              </SelectContent>
+            </Select>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+
+      {/* Enabled Switch */}
+      <FormField
+        control={form.control}
+        name="enabled"
+        render={({ field }) => (
+          <FormItem className="flex flex-row items-center justify-between">
+            <div className="space-y-0.5">
+              <FormLabel>Enable Monitor</FormLabel>
+              <FormDescription>
+                {isEdit
+                  ? "Monitor will run at the configured interval when enabled"
+                  : "Start monitoring immediately after creation"}
+              </FormDescription>
+            </div>
+            <FormControl>
+              <Switch checked={field.value} onCheckedChange={field.onChange} />
+            </FormControl>
+          </FormItem>
+        )}
+      />
+    </div>
+  );
+}
