+++ client/src/app/uptime/components/dialog/MonitorDialog.tsx
@@ -1,0 +1,261 @@
+import React, { useEffect } from "react";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+import { Button } from "@/components/ui/button";
+import { useCreateMonitor, useUpdateMonitor, UptimeMonitor } from "@/api/uptime/monitors";
+import { toast } from "sonner";
+import { Loader2 } from "lucide-react";
+import {
+  createMonitorSchema,
+  updateMonitorSchema,
+} from "../monitorSchemas";
+import { Form } from "@/components/ui/form";
+import { authClient } from "../../../../lib/auth";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/basic-tabs";
+import { GeneralTab } from "./GeneralTab";
+import { AdvancedTab } from "./AdvancedTab";
+import { RegionsTab } from "./RegionsTab";
+import { NotificationsTab } from "./NotificationsTab";
+
+interface MonitorDialogProps {
+  monitor?: UptimeMonitor;
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+}
+
+export function MonitorDialog({ monitor, open, onOpenChange }: MonitorDialogProps) {
+  const { data: activeOrganization, isPending } = authClient.useActiveOrganization();
+  const createMonitor = useCreateMonitor();
+  const updateMonitor = useUpdateMonitor();
+  const isEdit = !!monitor;
+
+  const form = useForm<any>({
+    resolver: zodResolver(isEdit ? updateMonitorSchema : createMonitorSchema),
+    defaultValues: isEdit
+      ? {
+          name: monitor.name,
+          intervalSeconds: monitor.intervalSeconds,
+          enabled: monitor.enabled,
+          httpConfig: monitor.httpConfig
+            ? {
+                url: monitor.httpConfig.url,
+                method: monitor.httpConfig.method as any,
+                headers: monitor.httpConfig.headers,
+                body: monitor.httpConfig.body,
+                auth: monitor.httpConfig.auth,
+                followRedirects: monitor.httpConfig.followRedirects ?? true,
+                timeoutMs: monitor.httpConfig.timeoutMs ?? 30000,
+                ipVersion: monitor.httpConfig.ipVersion ?? "any",
+                userAgent: monitor.httpConfig.userAgent,
+              }
+            : undefined,
+          tcpConfig: monitor.tcpConfig
+            ? {
+                host: monitor.tcpConfig.host,
+                port: monitor.tcpConfig.port,
+                timeoutMs: monitor.tcpConfig.timeoutMs ?? 30000,
+              }
+            : undefined,
+          validationRules: monitor.validationRules || [],
+          regions: monitor.regions || ["local"],
+        }
+      : {
+          organizationId: activeOrganization?.id || "",
+          name: "",
+          monitorType: "http" as const,
+          intervalSeconds: 300,
+          enabled: true,
+          httpConfig: {
+            url: "",
+            method: "GET" as const,
+            followRedirects: true,
+            timeoutMs: 30000,
+            ipVersion: "any" as const,
+          },
+          validationRules: [],
+          regions: ["local"],
+        },
+  });
+
+  const monitorType = isEdit ? monitor.monitorType : form.watch("monitorType");
+
+  const onSubmit = async (data: any) => {
+    console.log("Form submitted with data:", data);
+    console.log("Is edit mode:", isEdit);
+    
+    try {
+      if (isEdit) {
+        console.log("Updating monitor with ID:", monitor.id);
+        await updateMonitor.mutateAsync({
+          monitorId: monitor.id,
+          data,
+        });
+        toast.success("Monitor updated successfully");
+      } else {
+        await createMonitor.mutateAsync(data);
+        toast.success("Monitor created successfully");
+      }
+      onOpenChange(false);
+      form.reset();
+    } catch (error: any) {
+      console.error("Submit error:", error);
+      toast.error(error.response?.data?.error || `Failed to ${isEdit ? "update" : "create"} monitor`);
+    }
+  };
+
+  // Set organization ID when active organization changes (create mode only)
+  useEffect(() => {
+    if (!isEdit && activeOrganization?.id) {
+      form.setValue("organizationId", activeOrganization.id);
+    }
+  }, [activeOrganization, form, isEdit]);
+
+  // Reset form when dialog closes or monitor changes
+  useEffect(() => {
+    if (!open) {
+      form.reset();
+    } else if (isEdit && monitor) {
+      form.reset({
+        name: monitor.name,
+        intervalSeconds: monitor.intervalSeconds,
+        enabled: monitor.enabled,
+        httpConfig: monitor.httpConfig
+          ? {
+              url: monitor.httpConfig.url,
+              method: monitor.httpConfig.method as any,
+              headers: monitor.httpConfig.headers,
+              body: monitor.httpConfig.body,
+              auth: monitor.httpConfig.auth,
+              followRedirects: monitor.httpConfig.followRedirects ?? true,
+              timeoutMs: monitor.httpConfig.timeoutMs ?? 30000,
+              ipVersion: monitor.httpConfig.ipVersion ?? "any",
+              userAgent: monitor.httpConfig.userAgent,
+            }
+          : undefined,
+        tcpConfig: monitor.tcpConfig
+          ? {
+              host: monitor.tcpConfig.host,
+              port: monitor.tcpConfig.port,
+              timeoutMs: monitor.tcpConfig.timeoutMs ?? 30000,
+            }
+          : undefined,
+        validationRules: monitor.validationRules || [],
+        regions: monitor.regions || ["local"],
+      });
+    }
+  }, [monitor, open, form, isEdit]);
+
+  // Update form when monitor type changes (create mode only)
+  useEffect(() => {
+    if (!isEdit && form.watch) {
+      const subscription = form.watch((value, { name }) => {
+        if (name === "monitorType") {
+          if (value.monitorType === "tcp") {
+            form.setValue("httpConfig", undefined as any);
+            form.setValue("tcpConfig", {
+              host: "",
+              port: 80,
+              timeoutMs: 30000,
+            });
+          } else {
+            form.setValue("tcpConfig", undefined as any);
+            form.setValue("httpConfig", {
+              url: "",
+              method: "GET" as const,
+              followRedirects: true,
+              timeoutMs: 30000,
+              ipVersion: "any" as const,
+            });
+          }
+        }
+      });
+      return () => subscription.unsubscribe();
+    }
+  }, [form, isEdit]);
+
+  return (
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent className="max-w-2xl max-h-[90vh] overflow-y-auto">
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit, (errors) => {
+            console.log("Form validation errors:", errors);
+          })}>
+            <DialogHeader>
+              <DialogTitle>{isEdit ? "Edit" : "Create New"} Monitor</DialogTitle>
+              <DialogDescription>
+                {isEdit
+                  ? "Update the configuration for this monitor."
+                  : "Set up a new uptime monitor to track the availability of your endpoints."}
+              </DialogDescription>
+            </DialogHeader>
+
+            <Tabs defaultValue="general" className="mt-4">
+              <TabsList className="">
+                <TabsTrigger value="general">General</TabsTrigger>
+                <TabsTrigger value="advanced">Advanced</TabsTrigger>
+                <TabsTrigger value="regions">Regions</TabsTrigger>
+                <TabsTrigger value="notifications">Notifications</TabsTrigger>
+              </TabsList>
+
+              <TabsContent value="general" className="space-y-4 mt-4">
+                <GeneralTab 
+                  form={form} 
+                  monitor={monitor} 
+                  isEdit={isEdit} 
+                  monitorType={monitorType} 
+                />
+              </TabsContent>
+
+              <TabsContent value="advanced" className="space-y-4 mt-4">
+                <AdvancedTab 
+                  form={form} 
+                  monitorType={monitorType} 
+                />
+              </TabsContent>
+
+              <TabsContent value="regions" className="space-y-4 mt-4">
+                <RegionsTab />
+              </TabsContent>
+
+              <TabsContent value="notifications" className="space-y-4 mt-4">
+                <NotificationsTab />
+              </TabsContent>
+            </Tabs>
+
+            <DialogFooter className="mt-6">
+              <Button
+                type="button"
+                variant="outline"
+                onClick={() => onOpenChange(false)}
+                disabled={createMonitor.isPending || updateMonitor.isPending}
+              >
+                Cancel
+              </Button>
+              <Button
+                type="submit"
+                variant="success"
+                disabled={
+                  (isEdit ? updateMonitor.isPending : createMonitor.isPending) ||
+                  (!isEdit && (isPending || !activeOrganization))
+                }
+              >
+                {(createMonitor.isPending || updateMonitor.isPending) && (
+                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+                )}
+                {isEdit ? "Update" : "Create"} Monitor
+              </Button>
+            </DialogFooter>
+          </form>
+        </Form>
+      </DialogContent>
+    </Dialog>
+  );
+}
