+++ client/src/app/uptime/components/dialog/NotificationsTab.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function NotificationsTab() {
+  return (
+    <div className="text-neutral-500 text-sm">
+      Notification settings will be available in a future update.
+    </div>
+  );
+}
