+++ client/src/app/uptime/components/dialog/RegionsTab.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function RegionsTab() {
+  return (
+    <div className="text-neutral-500 text-sm">
+      Region selection will be available in a future update.
+    </div>
+  );
+}
