+++ client/src/app/uptime/components/dialog/index.ts
@@ -1,0 +1,1 @@
+export { MonitorDialog } from "./MonitorDialog";
